requirements.txt
aiogram==2.25.1
apscheduler==3.11.0
python-dotenv==1.1.1
